Use public getId instead of @TestApi getOriginalId
The PermissionsRequiredRouteProvider class was using the @TestApi
method MediaRoute2Info#getOriginalId in its onCreateSession callback
to make sure the selected route for the session was a known id. This
call is now getting a NoSuchMethodError exception, so this CL replaces
it with a call to the public MediaRoute2Info#getId.

The original version of this code didn't need to be using
getOriginalId in the first place, as that method is designed for
places where the package name prefix needs to be stripped off the
route id, but that is already taken care of in the machinery that
dispatches the onCreateSession callbacks to MediaRoute2ProviderService
instances.

Bug: 447033692
Flag: EXEMPT TEST_ONLY
Test: atest MediaRouter2HostSideTest
Change-Id: Id28c33e6b885abc68d3b0e1568377b5d4cf2ac01