Remove testGetServiceStateForSlotWithoutPermission
Since READ_PHONE_STATE is a required permission declared in the AndroidManifest of TelephonyManagerTest, it cannot be revoked at runtime for CTS testing.
If we set the App-Op mode to ignore, testGetServiceStateForSlotWithoutPermission will succeed, but the mode will have changed in an unexpected way, which may cause flaky failures in other tests.
The behavior of the getServiceStateForSlot API is already tested in testGetServiceStateForSlot, so we remove the test for the case where there is no permission.

Bug: 440943082
Test: atest TelephponyManagerTest
Flag: com.android.internal.telephony.flags.get_service_state_for_slot
Change-Id: Ia59b08e9b3fd0c41ccdd8300bee367e5223efe83