Replace EnsureCanAddUser with EnsureCanAddSecondaryUser
@EnsureCanAddUser checks to see whether the "overall max users" limit
has not been saturated. This is no longer reasonable behaviour, since
different users have different maxima.

Indeed, @EnsureCanAddUser is ambiguous, since being able to add one type
of user doesn't mean you can necessarily add another type of user. This
is especially true now that the max number of users is only on a
per-user-type basis.

In practice, @EnsureCanAddUser is currently almost always used to refer
to secondary users. So we should rename it to that.

For the few cases where it refers to profiles, it no longer actually
serves any purpose, as the "overall" max users no longer applies to
profiles anyway.

For the few cases where it refers to Guests, we leave the current
behaviour and mark with a TODO to fix this in the future, since it is
lower priority.

Bug: 442891661
Test: bedstead-multiuser-test
Flag: EXEMPT TEST_ONLY
Change-Id: I0634641c0808856daad7185f718e21fd9d77df54