Make appcompat flag mismatch an assumption failure in some tests
Some tests in the MediaRouter2HostSideTest suite use a helper method
on the host side of the test to set appcompat flags for the test app,
and the app-side of these tests then asserts that the flags are set to
the expected values. It turns out that helper method to set the flags
only works on eng and userdebug builds, and these tests are failing on
user builds.

As a temporary workaround, this CL changes these appcompat flag
assertions into assumption failures, so the tests won't fail on user
builds. We can then try and find a way to set the flags, or just skip
these tests, on user builds.

Bug: 447033692
Bug: 448042890
Test: atest MediaRouter2HostSideTest#testRestrictLocalNetworkCompatChange_notEnabled_routeIsFound,testRestrictLocalNetworkCompatChange_enabled_routeNotFound,testRestrictLocalNetworkCompatChange_enabled_routeFoundWhenWifiPermissionHeld (on a user build)
Flag: EXEMPT TEST_ONLY
Change-Id: I678aa118c7a7b0e9297b303cf134ec5286886cc8