Fix problem with appcompat flags on user builds
We were seeing failures of some CTS hostside tests that modify
appcompat flags, but only on user builds. It turns out that on user
builds the test app needs to be marked as debuggable in its
manifest. This CL does that, and removes a workaround we landed
earlier to work around the test failures.

Bug: 448491173
Test: atest MediaRouter2HostSideTest#testRestrictLocalNetworkCompatChange_enabled_routeNotFound,testRestrictLocalNetworkCompatChange_enabled_routeFoundWhenWifiPermissionHeld (on a user build)
Flag: EXEMPT TEST_ONLY
Change-Id: Ieae40844f77b35d1d10fd05f34313320f7291f05