ITS: Introduce PASS* test status for CameraITS.
PASS* test status will be applied only to each individual tests and does
not affect the scene status. It indicates if the test passes marginally
giving an early indication to the partners that the device might see
some issues with the test in future. It helps them look into any
possible fix early on.

When running the test individually, terminal will display the marginally
passing message ***Marginally passing test**. When running using the
run_all_tests tool, it will display the status as PASS* next to the test
and will also give the summary of total number of marginally passing
tests in the scene.

Bug: 407811574
Test: Tested locally
FLAG: TEST_ONLY
Change-Id: Idf5b6b2c706ac2760fab06ffde6bcacc0a642658