SonyDualshock4BluetoothTest: add debug logs
When running the battery test on Cuttlefish devices, we're sometimes
getting flakes where we timeout while waiting for the battery
capability. The logs on failure show a battery node in sysfs, but we
don't see the logs from InputReader that should appear when it's
processed and the input device re-initialized. Enable the debug logs for
SysfsNodeMonitor during this test, to help us narrow down the cause.

Bug: 437684441
Test: $ atest 'CtsHardwareTestCases:android.hardware.input.cts.tests.SonyDualshock4BluetoothTest#testBattery'
      and check that messages from SysfsNodeMonitor appear in logcat
Flag: EXEMPT TEST_ONLY
Change-Id: Id7962640983ae52d8096952a8b6d8046ef1c7c80