mediav2: modify producer in encoder color aspects test
The EncoderColorAspectsTest is currently using egl window as the
producer for the media codec component under test. The egl window
surface's dataspace has precendence over encoder's configuration via
MediaFormat. For MediaFormat's settings to take affect, the input
surface's dataspace has to be UNKNOWN. The current producer's dataspace
is not configurable. This change modifies the producer from egl window
to ImageWriter and configures the dataspace explicity to unknown and
verifies if the configured color aspects are taking affect

Bug: 417939517
Test: atest MctsMediaV2TestCases:EncoderColorAspectsTest
Test: atest CtsMediaV2TestCases:EncoderColorAspectsTest
Test: atest CtsVideoCodecTestCases
Flag: TEST_ONLY
Change-Id: I605158959dcf18003df2e39d581eb77773cf38da