Update version check for `testPropertiesIncludedInGetAllProperties`.
The test relies on a fix for large binder transactions in `getAllProperties` which was only included in 24Q4+ (AP4A - but since there is no version increment for that release, can only check B+). The test loads all of the DeviceConfig flags, so it randomly started failing due to the number of flags increasing over time. Rather than have the test require special circumstances (clear + no wifi), restrict to builds with the fix.

Change-Id: I26f75290ad097e9d6ef61e763cc3beeaf6e62d8b
Flag: EXEMPT TEST_ONLY
Bug: 445100894