Fix autofill test for automotive
On screens with small vertical dimensions, large insets, or scaling
the autofill picker bounds may be shifted due to lack of display space.
To address this, only compare the vertical alignment of the autofill
picker when it is touching the bottom of the screen. In automotive, we
additionally restrict where an app can be drawn, so the test needs to
calculate where the activity bounds end, instead of assuming the
activity height is the same as the display height.

Bug: 435657067
Test: atest android.autofillservice.cts.dropdown.VirtualContainerActivityCompatModeTest#testDatasetFiltering
Flag: EXEMPT TEST_ONLY
Change-Id: I2fee2ee0459923f7c64b542d90cfb908818d03bb