Set restriction on admin instead of user
In HSUM mode, CtsVerifier, run in user 0 and the created wifi config is
owned by user 0. Previously, the restriction was applied to user 10,
which was causing the 'share' feature to not be disabled.

Now, the restriction is applied to the device owner.

+ auto linter
+ extracted ButtonInfo objects to satisfy column limit

Bug: 448568511
Test: manual CtsVerifier > Device Owner Tests > Wifi configuration
lockdown > Disallow sharing admin configured WiFi confg
Flag: EXEMPT BUGFIX

Change-Id: I17355743c1053f495601c935372be9f164d84dbc