SELinuxNeverallowRule: Further truncate test names
As pointed out in b/446175448 comment #17, the previous commit
(47073729f77a28b7375d734834ea70a58fb1deef) failed to account
for the CTS prefix added to the test names, and as a result,
the tests were still exceeding the 512 byte limit. Account for
this extra data.

This extra truncation does not result in duplicated test names.

Fixes: 446175448
Flag: EXEMPT TEST_ONLY
Test: atest CtsSecurityHostTestCases
Change-Id: I12256242ed83137e170bfe87a26ead9852cba3c5