Skip IME Switcher Menu tests on wear
These tests are currently failing on wear with
config_preventImeStartupUnlessTextEditor when ran after some test that
requested showing the IME. The issue is the ImeContainer is not
re-parented correctly when the IME is destroyed with the config above
enabled, which is being investigated separately.

This skips the test for the time being on wear devices only, as the only
way to show the menu as an user on such devices is quite hard to
trigger, and does not get called in most cases.

Flag: EXEMPT TEST_ONLY
Bug: 406501026
Test: atest InputMethodManagerTest#testInputMethodPickerShownItems
  InputMethodManagerTest#testInputMethodPickerSwitchIme
Change-Id: I1aacf07a3b8ffd2bec757c5dc197a293978c84be