Add missing CTS test coverage in Telecom.
A smattering of missing test coverage.

Many were due to missing or incorrect "super" invocations.

Connection#onSeparate apparently was never implemented at all and looking
at the Telephony code would be a mistake to pipe through (it'll throw
exceptions).

TelecomAnalytics is system, and not used now that we have ww metrics.

VideoCall is interesting because we made an error in how we defined that
API to begin with.

Flag: EXEMPT TEST_ONLY
Test: CtsTelecomTestCases improvements.
Fixes: 449810148
Change-Id: I9642e9832be64f0b04ef928981ced5564d72771f