CtsMediaSessionTestCases: Generate ID by UUID instead of timestamp
Currently, the CtsMediaSessionTestCases is sometimes flaky due to
occasional collision of media session ID.

The collision happens because ID is generated based on timestamp ms,
and multiple calls to the function may happen in the same millisecond.

Change the ID generation to be based on Java's UUID, which should
minimize collision.

Test: atest android.media.session.cts.MediaSession2ServiceTest
Bug: 434605250
Flag: TEST_ONLY
Change-Id: I6527549d6932b0092928263153343fd76c09f1a2