Add test for Aconfig API in isolated process.
This change introduces a new test case to verify that the Aconfig API can be used within an Android isolated process.

A new service, `AconfigIsolatedProcessService`, is added with `android:isolatedProcess="true"`.

The test binds to this service and calls a method on it to read an Aconfig flag. The test asserts that the flag reading operation succeeds without crashing in the isolated process.

Change-Id: I2a96c68a221d8d405233ef01f5adaeed04bd4cac
Bug: 375154337
Flag: android.provider.flags.new_storage_public_api
Test: cts verified passed