Use another package for indirectActivityLaunch test
The previous fix worked to fix the test for most targets, but on wear
(cf_gwear_x86) the test flakes when it's run as part of the module.
Using package4 fixes the issue by avoiding using an activity in the
instrumentation process (and still works on other targets).

Bug: 445649158
Test: testRequestPinShortcut_indirectActivityLaunch on cf_gwear_x86
Flag: EXEMPT BUGFIX
Change-Id: I6b36b39ecd01b583e1ae31acab161a3f2507b2db