Use EthernetManager to check for Ethernet support
The PackageManager.hasSystemFeature(PackageManager.FEATURE_ETHERNET)
check is not always reliable for determining if a device truly supports
Ethernet. Some devices may not report this feature flag correctly,
leading to flaky test results in the JobScheduler CTS tests. A more
robust method is to directly check for the presence of the
EthernetManager system service. If the service is available, it's a
strong indicator that the device has Ethernet capabilities.

Test: atest CtsJobSchedulerTestCases
Bug: 449668748
Flag: EXEMPT BUGFIX
Change-Id: I1c12675f31adc4d73c24797833fe068de6c7f287