Use try-with in ProvisioningTest
The current problem with try-finally approach is that the exceptions
thrown from the try block get lost if the finally block throws more
exceptions.

Use try-with instead and set up a Resource that will get closed and
clean up the state automatically once the try block finishes the
execution.
Try-with also nicely combines the exceptions from the try block and from
the Resource.close method and no information gets lost.

Bug: 324222685
Test: btest android.devicepolicy.cts.ProvisioningTest*
Test: atest CtsDevicePolicyTestCases  --test-filter  "android.devicepolicy.cts.ProvisioningTest#*"
Flag: TEST_ONLY
Change-Id: I9b78a86e033e3081fbcd2d81675584b0353f68da