Use uinput touchscreen for injecting touches
This is the correct way to emulate touch. It ensures that the test is
hermetic. The injection of input events is deprecated and should no
longer be used.

While updating this test to use uinput-based injection, I noticed that
several tests started failing. This exposed the totally broken logic of
"mClicked" - the variable was supposed to be tracking the clicks on the
view. However, there was no proper waiting for this variable to change -
instead, the test was doing other incorrect synchronization methods in
order to try to ensure that "mClicked" has been updated.

In the new approach, a new class is introduced - ClickableButton. This
button stores the clicks directly inside the class, in a thread-safe
queue. We can then precisely time the asserts for waiting to receive
clicks, and also have better control over the negative asserts that the
click was not receive. This approach also forces the user to ensure that
all clicks are consumed (if there was a previous successful click, it
must be checked before the test could check that 'no click occurred'.

Bug: 448854710
Test: atest CtsSurfaceControlTests:SurfaceControlViewHostTests
Test: atest android.view.surfacecontrol.cts.SurfaceControlViewHostTests#testDisableInputTouch --iterations 3
Test: atest android.view.surfacecontrol.cts.SurfaceControlViewHostTests#testEmbeddedViewReceivesInput --iterations 3
Test: atest android.view.surfacecontrol.cts.SurfaceControlViewHostTests#testEmbeddedViewReleases --iterations 3
Flag: EXEMPT TEST_ONLY
Change-Id: I6382e8fa0de708f9e14fd126e7bccbc6e3f65e69