mediav2: fix race issue in ImageSurface implementation
The getImage() of ImageSurface.java dequeues image from the consumer
queue but does not close it. It is sent to the application for analysis.
The onImageAvailable() callback of ImageSurface.java, sees the consumer
queue to have vacant slots tries to add next image from image reader.
This is done by calling acquireNextImage(). But as the earlier dequeued
image is not actually closed, the image reader still owns it. If the
number of images in image reader match the max supported count, any call
to acquireNextImage() will raise Illegal state exception citing max
images has already been acquired. Before this call is made Image.close()
needs to be called. This change handles the same.

Bug: 452423515
Bug: 400189813
Bug: 374531092
Bug: 326010249
Test: atest MctsMediaV2TestCases:DecoderPushBlankBufferOnStopTest
Test: atest MctsMediaV2TestCases:CodecDecoderDetachedSurfaceTest
Test: atest CtsMediaV2TestCases:CodecDecoderDetachedSurfaceTest
Test: atest CtsMediaPerformanceClassTestCases:Av1FilmGrainValidationTest
Flag: TEST_ONLY
Change-Id: I5a715eef9b67376a561f2f35eeea8ebd68da02a2