ITS: Add the white balance check for JCA captures and preview.
This cl contains changes to get the RGB image from UltraHDR image and compare this RGB image with the JCA snapshot(preview which is not UltraHDR). The dynamic range patch cells are extracted and compared with each other to get the white balance difference.
if the diff is not within the threshold the test will fail. It also alerts if the test is passing marginally.

Bug: 427212603
Test: Tested locally
FLAG: TEST_ONLY
Change-Id: I2ca0faabaecd74dc02f51fdfea6f822430bdc5e9