Fixes ScreenCaptureTest in desktop mode
TestActivity was launched as a window in desktop mode. It has a
caption bar and rounded corners that can mess up the matched pixel
calculation in the screen capture test. This change crops out the
caption bar and rounded corners of the screenshot to fix the test.

Flag: EXEMPT TEST_ONLY
Test: atest ScreenCaptureTest
Bug: 452059112

Change-Id: I6899dea34a4b3fa25eca449f7a23b25821fdafa7