Wait for Activity resumed after split screen start
When launching an activity in split screen together with an existing
activity, the new activity might get re-launched. If the IME show
request is sent too early, before this re-launch, it might fail,
resulting in the IME being hidden.

This fixes the issue by waiting for all transitions to be finished on
the display, and for the new activity to be resumed.

Flag: EXEMPT TEST_ONLY
Bug: 451594103
Test: atest MultiUserMockImeTest#testHidingKeyboardInSplitScreenWithCrossProfileAppsCanShowKeyboardAgain
Change-Id: I995b7d1db79f28d877f2dc7ca60625d10b56f23f