Make graphics palette tests robust
SystemPaletteTest and DynamicColorsTest would time out and crash on Wear OS. The tests wait for an ACTION_CONFIGURATION_CHANGED broadcast
to confirm a theme change, but Wear OS does not consistently fire
this intent for all valid theme styles.

To fix this, we now use `assumeTrue(isSupportedStyle)` method in the shared BasePaletteTest to skip with unsupported Styles instead of crashing.

Apart from that test was failing in older HSUM build that do not have yet contrast fixed for these systems (ag/34761403). Re renamed `isSupportedDevice` to `isDynamicColorSupported` and check for the fix, HSUM and platform there.

Bug: 450143882
Bug: 448308903
Bug: 447136622
Bug: 440211538
Test: SystemPaletteTest
Flag: EXEMPT TEST_ONLY
Change-Id: I2634ec667e90e39854e23b20495bc5b58cc202b0