WebView: do not run CtsSdkSandboxWebkitTestCases on high-roi
No change to OS behavior. This updates the high-roi config files to
remove CtsSdkSandboxWebkitTestCases. Instead, this configures the bots
to run CtsWebkitTestCases, which are the same test cases but without the
overhead of enabling the SDK sandbox runtime.

The net effect of this change is that we should be running fewer tests
now on the high-roi bot, and the tests that we do still run are more
reliable because these won't need to create the SDK sandbox environment,
which we've seen to sometimes cause flakes.

Bug: 453666068
Test: N/A
Flag: TEST_ONLY
Change-Id: Ia1d87f1123d291e353fa0a48856e063b4490340f