Fix LocalVoicemailTest cases which are flaky/broken.
A change was made to the `MANAGED_DEFAULT_ACCOUNT_1` phone account
that is being used here to add `CAPABILITY_SIM_SUBSCRIPTION`; this
happened while my code was checked out but not rebased.  As a result
the phone account has to be run from the test runner which has shell
permission and can actually register the account.

Fixes: 452909427
Flag: com.android.server.telecom.flags.local_voicemail
Test: This is a test fix
Change-Id: I3c1d74124a7869a4c8fa66aa9ed9e79faad5acf0