Validate IllegalArgumentException thrown via setProperty
If the IllegalArgumentException is thrown because CarService received
StatusCode.INVALID_ARG from VHAL, then it is valid. All other
instances of IllegalArgumentException are not allowed as they are caused
by failed preconditions. The reason INVALID_ARG is allowed is because
there is no interface to define supported writable values and certain
values may be read-only.

Flag: EXEMPT TEST_ONLY
Bug: 447031408
Test: atest CtsCarTestCases:CarPropertyManagerTest
Change-Id: I3b764e814728359d39521902878b6e8056bada41