Hide IME at the end of split screen test
This test was added in [1], which ends with the IME visible in split
screen. During teardown we update the IME Control Target from the
RemoteInsetsControlTarget (RICT) to the activity, which leaves the IME
in the requestedVisibleTypes of the RICT.

This can lead to unexpected IME show requests later, for example in
cases like testRemovedImeOverlayLayeringTargetDoesNotAffectImeVisibility
where the IME Input Target is null throughout the test (as the window
has FLAG_NOT_FOCUSABLE).

This adds an IME hide request at the end of the test to avoid test
isolation issues, while tracking a proper fix for the underlying issue
with the RICT.

  [1]: Iaa475b5f2dabeb6b8d26627f160b425971d89284

Flag: EXEMPT TEST_ONLY
Bug: 450970010
Test: atest MultiUserMockImeTest#testHidingKeyboardInSplitScreenWithCrossProfileAppsCanShowKeyboardAgain
  followed by FocusHandlingTest#testRemovedImeOverlayLayeringTargetDoesNotAffectImeVisibility
Change-Id: I5237b10582204258d2a8f8dad267a6f17a0198b0