Fix PreferenceActivityFlowPortraitTest for automotive
The scrollToAndGetTextObject() tries to scroll to the top before slowly
scrolling down. However, the scroll() api only scrolls up equal to the
scrollable's visible height. This means that if scrollable is more than
half way down, it will not see the header. To fix this, continually
"fling" upwards until no more flings are allowed, signaling that the
scrollable is at the top of its content.

Bug: 437796936
Test: atest android.preference.cts.PreferenceActivityFlowPortraitTest#listDialogPortraitTest
on automotive seahawk and seahawk_cw
FLAG: EXEMPT TEST_ONLY

Change-Id: I672fc4abce258ef0340f88e86f8ad904d0126b01