[ITS] allow feature_combination parallel testing on gen2 rigs
This is similar to parallel testing for tablet-based scenes. Users should use CameraITS/tools/find_gen2_boards.py to determine which ports correspond to which rig, based on initialization (plug-in) time. config.yml must be filled out with two new fields to specify rotator and lighting ports for each rig.

For now, this feature is limited to test_feature_combination.

Flag: TEST_ONLY
Bug: 448649768
Test: ran test_feature_combination on two rigs in parallel. PASS.
Change-Id: I36089c4f4f099dc60c909cb8ea8bb3fc2a342e76