Adjust ContextualSearch background activity test to work for desktop.
The ContextualSearchFromOutOfProcessActivity_backgroundActivity test
currently assumes that starting a new activity would move the previous
activity to the background. However, this doesn't always seem to be the
case on desktop.

To make the test work on desktop devices, make it to simulate a HOME key
press to more reliably send an activity to the background.

Bug: 452293419
Test: atest ContextualSearchManagerTest
Flag: EXEMPT TEST_ONLY

Adjust ContextualSearch background activity test for desktop.

The ContextualSearchFromOutOfProcessActivity_backgroundActivity test currently assumes that starting a new activity would move the previous activity to the background. However, this doesn't always seem to be the case on desktop.

To make the test work on desktop devices, make it to simulate a HOME key press to more reliably send an activity to the background.

Bug: 452293419
Test: atest ContextualSearchManagerTest
Flag: EXEMPT test_only
Change-Id: I4a34e5c2c67137d0688c10570408095f6ffcb6c6