Update WindowManager private flag sanitization test
Previously, the test was incorrectly checking LayoutParams to
test whether the flags were sanitized. This just returns the
privateFlags that were passed as an argument from the test. This updates
the verification logic to check dumpsys output for the window.

Further, the method for changing the private flags is now updated to
match the exploit this is trying to fix.

Bug: 455764565, 406243581
Flag: EXEMPT TEST_ONLY
Test: atest android.security.cts.BUG_406243581.BUG_406243581

Change-Id: I8e332b1f653eba09da732c696e4c0c324c43b352
Merged-In: I8e332b1f653eba09da732c696e4c0c324c43b352