AudioManagerTest: Fix assertCallChangesStreamVolume() with alias stream
If the stream type is aliased, e.g. STREAM_ASSISTANT aliased as
STREAM_MUSIC, when the volume is changed, the ACTION_VOLUME_CHANGED
intent will be sent with the aliasStream in extra, instead of the
original stream type.

See
http://ac/frameworks/base/services/core/java/com/android/server/audio/AudioService.java?l=2934
for stream alias setting.

Note that the permission API call in testAssistantVolume() is updated
to the recommended one, to work with
assertCallChangesStreamVolume().
See go/bedstead-permissions .

Bug: 435563613
Bug: 383646023
Test: lunch cf_x86_64_desktop-trunk_staging-userdebug (also tested on cf_x86_64_phone-trunk_staging-userdebug)
Test: m && cvd create
Test: atest android.media.audio.cts.AudioManagerTest
Flag: TEST_ONLY
Change-Id: I1dda5bf0e78f0fccc9691ed02d42d333b316a422