mediav2: enable encoder output comparison across runs
This change saves encoder output to a buffer for each iteration and
verifies if they are bit-exact across all iterations. If they are not,
the results are logged.
TODO: If the results are different, then compare them for perceptual
similarity. If they are perceptually different, then raise
AssertionError.

Bug: 149027258
Test: atest MctsMediaV2TestCases:CodecEncoderTest
Test: atest CtsMediaV2TestCases:CodecEncoderTest
Test: atest MctsMediaV2TestCases:CodecEncoderSurfaceTest
Test: atest CtsMediaV2TestCases:CodecEncoderSurfaceTest
Flag: TEST_ONLY
Change-Id: I554ec861635e24ce99837a3b76e3fb3300801a05