[cts] fix archiveApp_dataIsKept with new cache file
To compensate for the different code paths encountered after package archival, the cache file created before the archival must be larger than the cache directory nodes size after the archival. This CL adds the creation of a new cache file of a certain size that is written when the test app is launched.

BUG: 453857978
Test: abtd
FLAG: TEST_ONLY
Change-Id: I1fd6482f4c940e5b9857500079499f7235e1f76a