CTS test for stopAndKill api behavior during app update
When flag is enabled, on package update, the app is stopped. We verify
this by having the app create a file which we can then check after the
update.

The test also verifies that it works with multiple packages being
updated at the same time, i.e, all packages are stopped. The goal is
to ensure we are not blocking the install thread by waiting on it.
We should be waiting in parallel.

Bug: 454290193
Flag: com.android.window.flags.enable_app_restart_after_update
Test: atest CtsStopAndKillHostTest
Change-Id: If0e3e394aa44b7227257fdea37d3e22804bcfab0