Fix flaky test due to shortened max animation duration
The test becomes flaky due to animation is too fast to become opaque before touch event happens. This test is to confirm if the bottom window is not obscured, it can be touched through. The test animation was setup to gradually become opaque as what a real app would do. But for testing this effect, we should simply keep the animation semi transparent (below threshold) from beginning to end.

Bug: 454393623
Test: WindowUntrustedTouchTest
Flag: EXEMPT TEST_ONLY
Change-Id: I0784eae09be16cd172c6dc1659b0ec7ccae8ceda