Resolve CTS test failure for testSetAudioRouteOnCallEndpointRequestedPropagation.
This fixes a CTS test failure that was occuring when 1 or less audio routes were available this test was failing in the finalyy block. I moved the check for how many available aduio routes there are outside of the try block so that if not enough audio routes are available, we immeadiately exit the test without entering the tray/catch/finally block at all.

Test: atest ThirdPartyInCallServiceAppOpsPermissionTest
Fixes: 457438992
Flag: android.telecom.flags.call_endpoint_requested_api
Change-Id: I7196a440d2ab531390ec10cd5111d1549ac14ab1