Add check to dismissErrorDialogIfPresent to avoid NPE
The method dismissErrorDialogIfPresent is designed to help tear down the
test cases in child test classes. When mShouldTestTelecom is false,
mInstrumentation will not be initialized and will result in
NullPointerExeption. This change add a check to avoid NPE in
CallSequencingManagedHoldRestrictionTest, CallSequencingMmiTest and
other future test cases that will use dismissErrorDialogIfPresent
method.

Bug: 428750183
Test: Test end correctly on devices that don't need to test telecom
Flag: EXEMPT TEST_ONLY
Change-Id: I04ab78179f69ec4603f3db41c222a8860a34aace