cts: Fix flaky AudioEffectTest#test4_0setEnabledLowerPriority
This test fails intermittently due to a race condition. It asserts
the control status of a low-priority effect immediately after
enabling a high-priority one, without waiting for the asynchronous
state change from AudioFlinger to reach the client.

This fix uses PollingCheck to wait for the control status to be
correctly updated before running the assertion.

Flag: TEST_ONLY
Bug: b/390846197
Test: atest android.media.audio.cts.AudioEffectTest#test4_0setEnabledLowerPriority

Change-Id: I47c96ae25544706531e25465400320e1cd9f737f