Fix flakiness in testIncomingCallDuringLocalVoicemailCall.
The test as originally written assumed that local VM will continue when
there is a new incoming call. It turns out my dev testing was completed
with the phone on silent to avoid annoying ringing.  When the phone is
not on silent the audio mode changes to MODE_RINGTONE when the ringtone
plays, so we HAVE to disconnect the local vm call in that case.  I have
split this out into two tests -- one assuming the phone is in silent mode
in which case the local vm call can process until the ringing call is
answered.  In the other case, a ringtone plays, which stops local
voicemail.

Test: Update and run localvoicemailtest.
Flag: android.telecom.flags.local_voicemail
Fixes: 458040987
Change-Id: If1556ed5408d42ab8ed17ba192d5a57512abd998