[cts] Ensure cache file is fully allocated
The method was previously creating a sparse file by just setting the length. This could lead to issues where the file doesn't actually occupy the intended space on disk. This change updates the method to write data to the file in chunks, ensuring that the file is fully allocated on disk. It also adds a log message to confirm the successful creation of the file.

BUG: 453857978
Test: atest and manually
FLAG: TEST_ONLY

Change-Id: I44fa884c39cb978075f61aeaa346e32b7ef6c2b3