Add test for PCC AppOps.
Verify that when setting an app op for an app uid, by default it is
inherited by a PCC uid that is defined by that app uid.

Not all app ops will follow this policy - we can add additional tests
when we have app ops where we don't want this inheritance to happen.

Bug: 437280042
Test: atest PccProcessTest
Flag: android.app.privatecompute.flags.enable_pcc_framework_support

Change-Id: I6943e4e5906ba639dd65175a720202ba382dd54c