Add fix orientation for IME screen off-on tests
This adds a fixed orientation session for IME tests that verify
behaviour by turning the screen off and on again, as the lockscreen
would modify the orientation to portrait. If this was not already
portrait, the configuration change could lead to undesired changes in
the TestActivity.

Flag: EXEMPT TEST_ONLY
Bug: 454660946
Test: atest ImeSwitchingTest#testImeSwitchingWithoutWindowFocusAfterDisplayOffOn
  KeyboardVisibilityControlTest#testHideImeAfterBackPressed_ScreenOffOn
Change-Id: I0a7760300eeffe83433a7337768655dbf06f312a