FrameRateOverride: Wait longer around modesets
On certain platforms even modesets for internal panels can take a while.
This leads to a delay in propogating actual refresh rate change
confirmations. As the beginning of some of the frame rate override tests
potentially invoke a modeset at the start of the test, it is important
to provide the extra grace period.

Additionally, SF internals might take some time to calibrate to the new
refresh rate due to abstractions like the vsync predictor. To support
that, take some extra time even after finding the target refresh rate to
make sure things are still working as intended.

Bug: 452259596
Test: atest FrameRateOverrideCtsActivity
Flag: EXEMPT BUGFIX
Change-Id: I964f931578584b0fdfbf135cafde044f07a1b168