Revert CtsAutoFillServiceTestCases resolution density workaround
9d979053e95a76e36097f7d322ddb79af6fb9f5e changes the display resolution and the pixel density before starting the test android.autofillservice.cts.dropdown.LoginActivityTest#testSwitchInputMethod_noNewFillRequest
to typical values for a mobile device with a vertical display.

This change should be reverted as it is more of a workaround than a proper fix and leads to problems for devices with a horizontal display.
It also lacks proper error handling to reset the settings at the end of the test in the event of an error.

Test: CtsAutoFillServiceTestCases on Google Pixel 10 Pro, Pixel tablet, Samsung S25+ like AOSP emulators
Bug: 456247278
Cherrypick-From: https://android-review.googlesource.com/q/commit:71726a976666a7809d2f9c19379c5d6d82737fd8
Merged-In: Ibcf1b7fd2edb43319f8a8704b0c4b5b4da9d3efb
Change-Id: Ibcf1b7fd2edb43319f8a8704b0c4b5b4da9d3efb