mediav2: update encoder test to choose encode config dynamically
CodecEncoderTest is intended to run for all encoders present on the
device. But sometimes the configuration that was chosen is not supported
by the component under test resulting the test to get skipped. This
change updates the test to select encoding configuration based on the
advertised capabilities

Bug: 457771555
Test: atest MctsMediaV2TestCases:CodecEncoderTest
Test: atest CtsMediaV2TestCases:CodecEncoderTest
Flag: TEST_ONLY

Change-Id: I5436047f699b803abfa110d6f861fa420a747216