Add debug logging for broadcast receiver timeouts.
Adds more debugging information to PackageBroadcastReceiver to
help investigate test failures. When a timeout occurs while
waiting for a broadcast, the status of it in the broadcast queue
is now dumped to the log. This will help in getting a better
understanding of the issue when the broadcast is not received
in time.

Bug: 452757457
Test: atest android.content.pm.cts.PackageManagerShellCommandMultiUserTest#testUpdateMimeGroup_changed
Flag: TEST_ONLY
Change-Id: I13bf10c0b7080b078b15529a3a058950262a8300