Wait for activity to stabilize in KeyboardVisibilityControlTest
The test testIMEVisibleInSplitScreenWithWindowInsetsApi is currently
flaky. We are observing that the coordinates that are used during touch
injection may be off. This could happen if the launched activity hasn't
finished launching. As a speculative fix, ensure that the geometry has
stabilized before we use the activity's coordinates to determine the
location of the injected tap.

Note: while running the test locally, I found that testShowHideKeyboardOnWebView
failed.

Bug: 456019015
Test: atest KeyboardVisibilityControlTest
Flag: EXEMPT TEST_ONLY
Change-Id: Id4eec2e5c069b1371eecbae0a2a95ebc1db9e0d4