Test fix: restricted profile parent may not be sys
The CTS test currently checks that the restricted profile's parent must
be the System user. However, on HSUM that is incorrect. So we update
the test to instead check that the parent is indeed who we expect it to
be and is an Admin (with no requirement that it be the system user).

Test: atest CtsMultiUserTestCases:android.multiuser.cts.UserManagerTest#testRestrictedUser
Bug: 454378716
Flag: EXEMPT BUG_FIX
Change-Id: Ida0d993c7885c33f0671e9a1e1a29c2d5491782a