Update CarWatchdogHostTest to send SHUTDOWN_PREPARE command instead of reboot command
VHAL is not required to respond to the shutdown request from carservice,
which causes some tests to occasionally fail if they try to shutdown the
device. Update the implementation to use a shutdown-prepare and
shutdown-enter command that can't be ignored, allowing the test to pass.
This change was already merged for ATS, but merge for CTS as well.

Test: atest CarWatchdogHostTest
Flag: EXEMPT TEST_ONLY
Fixes: 460236014
Change-Id: I8f3774cbc7d7c4d1eb62e6b78cec325a391dfdac