fix: Ignore PiP test on TV which doesn't report PiP window changes.
The internal Android TV team is investigating fixing this PiP issue
in b/434237438, but this is not something that an OEM can resolve so
we should not block OEM validation on this missing TV-platform behavior.

See commit 1ba5892365893e59bfbc1d232a6f6cd47ce70f34
which skipped a similar test for the same reason.

Bug: 434237438
Bug: 455454125
Test: n/a (ignoring existing test)
Flag: EXEMPT TEST_ONLY
Change-Id: I46c52f20be0a38aa3b17ed05a6fba6731d085d09